requirements.txt
# Core framework
Flask==3.0.2

# Environment variable handling
python-dotenv==1.0.1

# Utilities
mysql-connector-python==8.3.0
